Added delete functionality in IOS with smooth animations

+++ androidApp/src/main/java/com/ps/translatepro/android/translate/presentation/TranslateScreen.kt
@@ -195,7 +195,7 @@ fun TranslateScreen(
                     }
                     onEvent(TranslateEvent.SelectHistoryItem(historyItem))
                 }, onDelete = {
-                    onEvent(TranslateEvent.DeleteHistoryItem(historyItemId = historyItem.id))
+                    onEvent(TranslateEvent.DeleteHistoryItem(historyItemId = historyItem.id.toInt()))
 
                 }, modifier = Modifier.fillMaxWidth()
                 )
